Copied over cute wizards code

+++ engine/DraggableElement.cpp
@@ -10,9 +10,9 @@ public:
    DraggableElement(std::unique_ptr<UIElement> inner)
       :DecoratorUIElement(std::move(inner)), m_dragging(false)
    {
-      m_inner->registerMouseButton(Input::MouseLeft, Input::Press, 0, [&](MouseEvent e){onMouseDown(e.x, e.y);});
-      m_inner->registerMouseButton(Input::MouseLeft, Input::Release, 0, [&](MouseEvent e){onMouseUp(e.x, e.y);});
-      m_inner->registerMouseMove([&](MouseEvent e){onMouseMove(e.x, e.y);});
+      m_inner->registerMouseButton(Input::MouseLeft, Input::Press, 0, [&](MouseEvent e){onMouseDown((float)e.x, (float)e.y);});
+      m_inner->registerMouseButton(Input::MouseLeft, Input::Release, 0, [&](MouseEvent e){onMouseUp((float)e.x, (float)e.y);});
+      m_inner->registerMouseMove([&](MouseEvent e){onMouseMove((float)e.x, (float)e.y);});
    }
 
    void onMouseDown(float x, float y)
